Mitigation for upcoming burn dynamic memory allocations
WebGPU freaks out because different parts of a buffer are bound as read and read_write. So, for now make everything read_write.

This is a shame and will introduce needless barriers (and other missed optimizations?), but not sure what to do here atm. This might be a wgpu spec bug.