Allow training with alpha masks, more consistent background handling.
Currently, alpha masks are ignored and a background is composed into training images. This instead allows gradients to flow to force alpha to match the inputs. This gets weird though, as the code tries to composite some hard-coded background. This is done in a strange way, where it does composit but affect the alpha, which can break things. Instead just remove this background composition from the kernels - it can be done cheaply as a normal tensor op.

Also add a nice checkered background for alpha images.